Use external SD card filesystem instead of internal
The internal filesystem will be overwritten when a new firmware app for
the ODOID GO is loaded. Mount the SD card on boot to '/sd', change
directory there, and look for a 'boot.py' and 'main.py' to execute.

If no 'boot.py' is found, create one that appends '/sd/lib' to the
system path.